Using existing EndpointInfo class
